requirements.txt
streamlit
openai
google-generativeai
groq
python-dotenv
Pillow
pandas
audio-recorder-streamlit
SpeechRecognition
pydub
requests